ci: tox stages only clean up docker image if it exists

+++ Jenkinsfile
@@ -943,7 +943,9 @@ pipeline {
                                                                 }
                                                             }
                                                         } finally {
-                                                            sh "docker rmi ${image.id}"
+                                                            if (image){
+                                                                sh "docker rmi --force --no-prune ${image.id}"
+                                                            }
                                                         }
                                                     }
                                                 }
@@ -1035,7 +1037,9 @@ pipeline {
                                                                 bat "${tool(name: 'Default', type: 'git')} clean -dfx"
                                                             }
                                                         } finally{
-                                                            bat "docker rmi --force --no-prune ${image.id}"
+                                                            if (image){
+                                                                bat "docker rmi --force --no-prune ${image.id}"
+                                                            }
                                                         }
                                                      }
                                                  }
